+++ src/core/epoch.h
@@ -30,58 +30,16 @@
 // info@tripwire.org or www.tripwire.org.
 //
 ///////////////////////////////////////////////////////////////////////////////
-// timebomb.h
-#include "stdcore.h"
-#include "timebomb.h"
-#include <time.h>
-#include <iostream>
-#include "timeconvert.h"
-#include "corestrings.h"
+// epoch.h
+
+#ifndef __EPOCH_H
+#define __EPOCH_H
 
 ///////////////////////////////////////////////////////////////////////////////
-// TimeBombExploded() -- Call from main(). Prints out timebomb message and
+// CheckEpoch() -- Call from main(). Prints out timebomb message and
 //      returns true if main() should exit.
 //
-bool TimeBombExploded()
-{
-    struct tm time_struct;
+bool CheckEpoch();
 
-    /*
-    memset(&time_struct, 0, sizeof(time_struct));
-    time_struct.tm_mday = 25;
-    time_struct.tm_mon = 0;
-    time_struct.tm_year = 99;
-    int64 begin = cTimeUtil::DateToTime( &time_struct );
-
-    memset(&time_struct, 0, sizeof(time_struct));
-    time_struct.tm_mday = 1;
-    time_struct.tm_mon = 4;
-    time_struct.tm_year = 99;
-    int64 end = cTimeUtil::DateToTime( &time_struct );
-
-    int64 now = time(0);
-
-    if (now < begin || now > end)
-    {
-        std::cerr << "This beta version of Tripwire(R) has expired.\n";
-        return true;
-    }
-    */
-
-    // Many functions will fail as we approach the end of the epoch
-    // Rather than crashing, we will exit with a nice message
-    memset(&time_struct, 0, sizeof(time_struct));
-    time_struct.tm_mday = 1;
-    time_struct.tm_mon = 0;
-    time_struct.tm_year = 138;
-    int64 endoftime = cTimeUtil::DateToTime( &time_struct );
-
-    if (time(0) > endoftime)
-    {
-        TCERR << TSS_GetString(cCore, core::STR_ENDOFTIME) << std::endl;
-        return true;
-    }
-
-    return false;
-}
+#endif
