+++ src/siggen/siggenmain.cpp
@@ -43,7 +43,7 @@
 #include "core/cmdlineparser.h"
 #include "core/errorbucketimpl.h"
 #include "core/usernotifystdout.h"
-#include "core/timebomb.h"
+#include "core/epoch.h"
 #include "fs/fsstrings.h"
 #include "tw/twstrings.h"
 #include "tw/twutil.h"      // for cTWUtil::PrintErrorMsg
@@ -104,7 +104,7 @@ int __cdecl _tmain(int argc, const TCHAR** argv)
 {
     int ret = 0;
 
-    if (TimeBombExploded())
+    if (CheckEpoch())
         return 1;
 
     try
